Added Auto code and fixed int problems

+++ org/usfirst/frc/team435/robot/commands/DriveForward.java
@@ -9,9 +9,9 @@ import edu.wpi.first.wpilibj.command.Command;
  */
 public class DriveForward extends Command {
 
-	private int speed;
+	private double speed;
 	
-    public DriveForward(int speed, int time) {
+    public DriveForward(double speed, double time) {
         this.speed = speed;
         setTimeout(time);
         requires(Robot.driveTrain);
