+++ org/usfirst/frc/team435/robot/subsystems/DriveTrain.java
@@ -9,13 +9,13 @@ import edu.wpi.first.wpilibj.command.Subsystem;
  */
 public class DriveTrain extends Subsystem {
     
-    public void driveForward(int speed) {
+    public void driveForward(double speed) {
     	RobotMap.robotDrive.mecanumDrive_Cartesian(0, -speed, 0, 0);
     }
-    public void strafeLeft(int speed) {
+    public void strafeLeft(double speed) {
     	RobotMap.robotDrive.mecanumDrive_Cartesian(speed, 0, 0, 0);
     }
-    public void turnLeft(int speed) {
+    public void turnLeft(double speed) {
     	RobotMap.robotDrive.mecanumDrive_Cartesian(0, 0, speed, 0);
     }
     public void initDefaultCommand() {
